backend/requirements.txt
fastapi
uvicorn
google-genai
python-multipart
httpx
pytest
sqlalchemy[asyncio]
aiosqlite
